Refactor go modules
