+++ plugins/modules/src/flavor_info/flavor_info.go
@@ -14,7 +14,7 @@
  * Copyright 2025 Red Hat, Inc.
  *
  */
-package main
+package flavor_info
 
 import (
 	"context"
@@ -66,7 +66,7 @@ func returnResponse(responseBody Response) {
 	}
 }
 
-func main() {
+func Run() {
 	var response Response
 	if len(os.Args) != 2 {
 		response.Msg = "No argument file provided"
